\name{ae-package}
\alias{ae-package}
\alias{ae}
\docType{package}
\title{
\packageTitle{ae}
}
\description{
\packageDescription{ae}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ae}
\packageIndices{ae}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{ae}

Maintainer: \packageMaintainer{ae}
}
\references{
-
}

\keyword{ package }
\seealso{
\url{https://www.researchgate.net/publication/252568552_A_new_measure_of_brand_attitudinal_equity_based_on_the_Zipf_distribution}

}
\examples{
varstocases <- v2c(df, "respid", "q1", "q2")
ae <- ae(data = a, id = "respid", "relevance", "closeness")

}
