% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{df}
\alias{df}
\title{Example data}
\format{A data frame with 10 rows and 11 variables:
\describe{
  \item{respid}{id}
  \item{q1_1}{Relevance Brand 1}
  \item{q1_2}{Relevance Brand 2}
  \item{q1_3}{Relevance Brand 3}
  \item{q1_4}{Relevance Brand 4}
  \item{q1_5}{Relevance Brand 5}
  \item{q2_1}{Closeness Brand 1}
  \item{q2_2}{Closeness Brand 2}
  \item{q2_3}{Closeness Brand 3}
  \item{q2_4}{Closeness Brand 4}
  \item{q2_5}{Closeness Brand 5}
}}
\usage{
df
}
\description{
A dataframe that contains dummy data in wide format for computing Attitudinal Equity
}
\keyword{datasets}
