% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v2c.R
\name{v2c}
\alias{v2c}
\title{VarsToCases}
\usage{
v2c(data, id, relevance_root_q, closeness_root_q)
}
\arguments{
\item{data}{A data frame.}

\item{id}{name of id variable, within "".}

\item{relevance_root_q}{question root for relevance variables, format questionroot_brandid, within "".}

\item{closeness_root_q}{question root for closeness variables, format questionroot_brandid, within "".}
}
\value{
Returns transposed data frame (long format).
}
\description{
Function is intended to work as SPSSs VarsToCases
}
